Add "transformation functions" to $computed, tests
Transformation functions are used when $computed is given an array of
functions instead of a single function. The functions are each called in
sequence with the result of the previous function (the first function
gets no argument). These allow a series of computations to be chained
(with extraction in between) without going through the $digest cycle.
Each step is free to use separate dependencies. If any dependencies in
the chain prompt a recalculation the recalculation will start as far
down the chain as possible.