Fix examples, add debug mode
The examples were looking at dist/ng-computed.js, because I forgot to
update them to look at just ../ng-computed.js, so I fixed that. I also
changed basic-example and direct-example to be a bit nicer, and to
report using the new debug stuff.

The debug stuff that I've added for now is just for reporting on
dependency tracking. It provides a new method,
$computed.dependencyGraph() which gives you a map: Expr -> WatchKey ->
Value, which lets you look up what values a computed expression depends
on, what kind of watch they are and what value they had the last time
they were used. This information can be useful to see why something isn't
behaving as it might be expected to behave.